Move logic in util
