+++ core/src/collections.rs
@@ -1,5 +1,5 @@
-use crate::run_command::run_command_with_retry;
 use crate::stmt::EmptyStatement;
+use crate::util::run_command_with_retry;
 use crate::util::{
     is_match, table_type_filter_to_vec, to_name_regex, DISALLOWED_COLLECTION_NAMES,
     DISALLOWED_DB_NAMES,
